Cleanup stables module
