+++ tests/Feature/Admin/Stables/ViewStableBioPageSuccessConditionsTest.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace Tests\Feature\Admin\Stables;
+
+use Tests\TestCase;
+use App\Models\Stable;
+use Illuminate\Foundation\Testing\RefreshDatabase;
+
+/**
+ * @group stables
+ * @group admins
+ */
+class ViewStableBioPageSuccessConditionsTest extends TestCase
+{
+    use RefreshDatabase;
+
+    /** @test */
+    public function an_administrator_can_view_a_stable_profile()
+    {
+        $this->actAs('administrator');
+        $stable = factory(Stable::class)->create();
+
+        $response = $this->get(route('roster.stables.show', $stable));
+
+        $response->assertViewIs('stables.show');
+        $this->assertTrue($response->data('stable')->is($stable));
+    }
+}
